Contacts: Fix "Click to select group" button disappeared
"Click to select group" button disappeared after delete added
group and rotate screen in contact edit screen. The reason is
that mEntry.mAfter of LocalGroupsSelector will be set to null
when delete added group. Then entry.isVisible() in setValues()
of LabeledEditorView is false, so the visibility of
"Click to select group" button will be set to View.GONE.

Just remove the value of column in mEntry.mAfter instead that
set mEntry.mAfter of LocalGroupsSelector to null.

CRs-Fixed: 630120

Change-Id: Ia1bb5646f3148e7536ca3a0fe226d63bfdeaaf61